Introduced worker_shutdown_idle_delay.
The directive configures a delay before closing idle connections to be used
when gracefully shutting down worker processes. When the timer expires, nginx
will attempt to close any remaining idle connections.

The delay is only added for connections with the "shutdown_delay" flag, and
currently it's only keepalive HTTP/1.1 connections. The behavior is not changed
for protocols that have built-in graceful shutdown mechanisms, such as GOAWAY
frames in HTTP2 and HTTP3.

Although it's perfectly fine to close an HTTP/1.1 connection at any time
according to RFC 9112, it may still be useful to delay closing a connection,
wait for the next client request, and close the connection with a
"Connection: close" header to avoid unnecessary retries by clients. This is
especially important for environments with frequent nginx reloads and large
amounts of non-idempotent requests which are quite problematic for automatic
retries.

Should be used carefully to not delay configuration reloads too much (and thus
increase nginx resource usage), and ideally in combination with properly
configured clients:

client_idle_timeout < min(server_shutdown_delay, server_idle_timeout)